impl(storage): make the `OpenObject` builder public
